Bind PluginManager optionally
