+++ testing/trino-server-dev/etc/log.properties
@@ -8,7 +8,7 @@
 io.trino=INFO
 
 # show classpath for plugins
-io.trino.server.PluginManager=DEBUG
+io.trino.server.TrinoPluginManager=DEBUG
 
 # Maven plugin loading code
 com.ning.http.client=WARN
